+++ lib/Stomp/Comunication/LocalComunicationInterface.php
@@ -1,0 +1,8 @@
+<?php
+namespace Secuconnect\Client\STOMP\Comunication;
+use \Secuconnect\Client\STOMP\Client;
+interface LocalComunicationInterface
+{
+    public function getMsgToSendForStompController();
+    public function sendMsgToStompController($msg);
+}