+++ lib/Stomp/Comunication/ReceivedFrameControllerInterface.php
@@ -1,0 +1,6 @@
+<?php
+namespace Secuconnect\Client\STOMP\Comunication;
+interface ReceivedFrameControllerInterface
+{
+    public function process($frame);
+}