+++ lib/Stomp/Comunication/ReceivedFrameFileController.php
@@ -1,0 +1,12 @@
+<?php
+
+namespace Secuconnect\Client\STOMP\Comunication;
+use Secuconnect\Client\STOMP\Client;
+class ReceivedFrameFileController implements ReceivedFrameControllerInterface
+{  
+
+    public function process($frame) {
+        var_dump($frame);
+    }
+    
+}